refactor: DTO 네이밍 추가 수정
