+++ src/main/java/com/kdt/team04/domain/teams/teaminvitation/repository/CustomTeamInvitationRepository.java
@@ -4,7 +4,7 @@ import com.kdt.team04.common.PageDto;
 import com.kdt.team04.domain.teams.teaminvitation.dto.TeamInvitationCursor;
 import com.kdt.team04.domain.teams.teaminvitation.dto.response.TeamInvitationResponse;
 
-public interface TeamInvitationRepositoryCustom {
+public interface CustomTeamInvitationRepository {
 	PageDto.CursorResponse<TeamInvitationResponse, TeamInvitationCursor> getInvitations(
 		Long targetId, PageDto.TeamInvitationCursorPageRequest request);
 }
