+++ src/main/java/com/kdt/team04/domain/teams/teaminvitation/repository/TeamInvitationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.kdt.team04.domain.teams.teaminvitation.model.InvitationStatus;
 import com.kdt.team04.domain.teams.teaminvitation.model.entity.TeamInvitation;
 
-public interface TeamInvitationRepository extends JpaRepository<TeamInvitation, Long>, TeamInvitationRepositoryCustom {
+public interface TeamInvitationRepository extends JpaRepository<TeamInvitation, Long>, CustomTeamInvitationRepository {
 	boolean existsByTeamIdAndTargetIdAndStatus(Long teamId, Long targetId, InvitationStatus status);
 
 	Optional<TeamInvitation> findByTeamIdAndTargetId(Long teamId, Long targetId);
